PR: Optimize downloader (#62)
* Add options to upload/download CLI commands

This was especially helpful for benchmarking.

* Reduce download chunk size from 64MB to 4MB

Based on experiments, 4MB was the optimal download chunk size for any
number of threads.

* Update default chunk size

* Handle keyboard interrupt and exceptions

* Revert handling exceptions

The goal was to print out any exceptions and keep the command loop
running (good for interactive testing). At this time, I'm holding off on
this until I can handle it robustly while balancing subcommand
exceptions and keyboard events.