Fix missing chunk metadata when downloading
For downloads with no merging, each chunk is distinguished by two items:
name and offset. When submitting a file and splitting into chunks, we
create a future and assign it to a dictionary. Previously we used the
chunk name as the key but this caused multiple chunks to be assigned to
the same key. Thus, the chunk metadata was incorrect and in some cases
caused us to lose track of futures.