PR: Metadata improvements (#76)
* Split file into chunks during submission

Previously we split a file into chunks when running the client. This
made it difficult to verify during tests whether files were split
correctly.

* Rename parameter argument

* Rename nbytes to length

This was done to make the arguments clearer.

* Fix missing chunk metadata when downloading

For downloads with no merging, each chunk is distinguished by two items:
name and offset. When submitting a file and splitting into chunks, we
create a future and assign it to a dictionary. Previously we used the
chunk name as the key but this caused multiple chunks to be assigned to
the same key. Thus, the chunk metadata was incorrect and in some cases
caused us to lose track of futures.

* Add test

* Use callback, not polling, to signal end of task

I restructured the internal file/chunk state so that we could look at
state based on a given future (driven by the need of callbacks). This in
turn simplified the status check in _update().

* Fix undefined variable

* Add internal documentation

* Capture expected/actual bytes per chunk

* Update CLI default chunk size

* Fix test

* Return bytes written and exception string

Instead of raising an exception, we catch the exception and return it as
part of a tuple.

* Check chunk for exception or byte mismatch

We set the chunk as errored if either is true.

* Create unique temp directory for each transfer

Each chunk is placed in the same temporary directory now. A previous bug
put each chunk into its own unique temporary directory. This caused
collisions that affected file size.

* Enforce unique temp directory per file

* Increase uniqueness of temporary directory

* Ensure we are always appending at the offset

* Add retry logic to REST calls in flush

* Fix debug message

* Sort chunks by offset before merging

* Skip some exceptions when retrying

We skip the retry logic when we encounter missing files/directories or
bad permissions.

* Revert changed exception superclass

* Add path to exception message