PR: Actionable errors (#94)
* Removed unused file/chunk metadata

A file's start/stop time and a chunk's REST call retries is not
relevant.

* Remove unused imports

* Refactor debug and error messages

When debugging, the request and response is grouped respectively into a
single message for clearer logs. We also now record Azure operation and
parameters.

When an exception occurs, we record the response and body along with the
exception. This will happen with all HTTP status codes in the 4xx or 5xx
range.

* Tidy printed response content and limit if needed

* Check for valid response

Oddly enough, `if response` doesn't work, but `if response is not None`
does.