merge in the changes from master, keep synced.
There were some build process changes made in master that were not
applied to dev. Applying them to keep everything in sync for
development.

Address an issue with throttling on download

This should address an issue with download where, if throttled, less
than the requested data is returned. In this case, we should check to
ensure we got as much data as we wanted prior to returning.

slight updates to how total bytes is calculated