Update Retry logic to handle throttling scenarios
This change enables upload and download to more reliably complete
successfully in the face of intermittent server-side throttling by
enacting a back-off retry strategy. Note that this strategy is not
guaranteed to work every time. For example, if the server is under
extremely heavy load and continues to reject requests, the retry
strategy will eventually "give up" and return an error back to the user.